+++ cravat/cravat_class.py
@@ -23,6 +23,7 @@ import cravat.cravat_util as cu
 import collections
 import asyncio
 import sqlite3
+from .inout import CravatWriter
 
 cravat_cmd_parser = argparse.ArgumentParser(
     prog='cravat input_file_path_1 input_file_path_2 ...',
@@ -166,6 +167,7 @@ class Cravat (object):
         self.should_run_reporter = True
         self.pythonpath = sys.executable
         self.annotators = {}        
+        self.append_mode = False
         self.make_args_namespace(kwargs)
         self.get_logger()
         self.start_time = time.time()
@@ -426,9 +428,9 @@ class Cravat (object):
         dbpath = os.path.join(self.output_dir, self.run_name + '.sqlite')
         conn = sqlite3.connect(dbpath)
         cursor = conn.cursor()
-        q = 'create table if not exists smartfilters (name text, definition text)'
+        q = 'create table if not exists smartfilters (name text primary key, definition text)'
         cursor.execute(q)
-        ins_template = 'insert into smartfilters (name, definition) values (?, ?);'
+        ins_template = 'insert or replace into smartfilters (name, definition) values (?, ?);'
         for linfo in self.annotators.values():
             if linfo.smartfilters is not None:
                 mname = linfo.name
@@ -495,6 +497,10 @@ class Cravat (object):
             self.run_name = os.path.basename(self.inputs[0])
             if len(self.inputs) > 1:
                 self.run_name += '_and_'+str(len(self.inputs)-1)+'_files'
+        if self.inputs[0].endswith('.sqlite'):
+            self.append_mode = True  
+            if self.run_name.endswith('.sqlite'):
+                self.run_name = self.run_name[:-7]
         self.output_dir = self.args.output_dir
         if self.output_dir == None:
             self.output_dir = os.path.dirname(os.path.abspath(self.inputs[0]))
@@ -529,6 +535,8 @@ class Cravat (object):
             self.args.startat = 'aggregator'
         if 'postaggregator' in self.args.repeat and not 'aggregator' in self.args.repeat:
             self.args.repeat.append('aggregator')
+        if self.append_mode:
+            self.args.endat = 'aggregator'
         try:
             self.startlevel = self.runlevels[self.args.startat]
         except KeyError:
@@ -574,6 +582,55 @@ class Cravat (object):
             self.crg_present = True
         else:
             self.crg_present = False
+        
+        if self.append_mode:
+            self.regenerate_from_db()
+    
+    def regenerate_from_db (self):
+        dbpath = self.inputs[0]
+        db = sqlite3.connect(dbpath)
+        c = db.cursor()
+        # Variant
+        if not self.crv_present:
+            crv = CravatWriter(self.crvinput, columns=constants.crv_def)
+            crv.write_definition()
+        else:
+            crv = None
+        if not self.crx_present:
+            crx = CravatWriter(self.crxinput, columns=constants.crx_def)
+            crx.write_definition()
+        else:
+            crx = None
+        if crv or crx:
+            colnames = [x['name'] for x in constants.crx_def]
+            sel_cols = ', '.join(['base__'+x for x in colnames])
+            q = f'select {sel_cols} from variant'
+            c.execute(q)
+            for r in c:
+                rd = {x[0]:x[1] for x in zip(colnames,r)}
+                if crv:
+                    crv.write_data(rd)
+                if crx:
+                    crx.write_data(rd)
+            crv.close()
+            crx.close()
+            self.crv_present = True
+            self.crx_present = True
+        # Gene
+        if not self.crg_present:
+            crg = CravatWriter(self.crginput, columns=constants.crg_def)
+            crg.write_definition()
+            colnames = [x['name'] for x in constants.crg_def]
+            sel_cols = ', '.join(['base__'+x for x in colnames])
+            q = f'select {sel_cols} from gene'
+            c.execute(q)
+            for r in c:
+                rd = {x[0]:x[1] for x in zip(colnames,r)}
+                crg.write_data(rd)
+            crg.close()
+            self.crg_present = True
+        c.close()
+        db.close()
 
     def populate_secondary_annotators (self):
         secondaries = {}
@@ -682,6 +739,8 @@ class Cravat (object):
                '-n', self.run_name]
         if self.cleandb:
             cmd.append('-x')
+        if self.append_mode:
+            cmd.append('--append')
         if self.verbose:
             print(' '.join(cmd))
         self.update_status('Running {title} ({level})'.format(title='Aggregator', level='variant'))
@@ -698,6 +757,8 @@ class Cravat (object):
                '-d', self.output_dir, 
                '-l', 'gene',
                '-n', self.run_name]
+        if self.append_mode:
+            cmd.append('--append')
         if self.verbose:
             print(' '.join(cmd))
         self.update_status('Running {title} ({level})'.format(title='Aggregator', level='gene'))
@@ -707,35 +768,37 @@ class Cravat (object):
         print('finished in {0:.3f}s'.format(rtime))
 
         # Sample level
-        print('\t{0:30s}\t'.format('Samples'), end='', flush=True)
-        stime = time.time()
-        cmd = ['donotremove', 
-               '-i', self.output_dir,
-               '-d', self.output_dir, 
-               '-l', 'sample',
-               '-n', self.run_name]
-        if self.verbose:
-            print(' '.join(cmd))
-        self.update_status('Running {title} ({level})'.format(title='Aggregator', level='sample'))
-        s_aggregator = Aggregator(cmd, self.status_writer)
-        s_aggregator.run()
-        rtime = time.time() - stime
-        print('finished in {0:.3f}s'.format(rtime))
+        if not self.append_mode:
+            print('\t{0:30s}\t'.format('Samples'), end='', flush=True)
+            stime = time.time()
+            cmd = ['donotremove', 
+                '-i', self.output_dir,
+                '-d', self.output_dir, 
+                '-l', 'sample',
+                '-n', self.run_name]
+            if self.verbose:
+                print(' '.join(cmd))
+            self.update_status('Running {title} ({level})'.format(title='Aggregator', level='sample'))
+            s_aggregator = Aggregator(cmd, self.status_writer)
+            s_aggregator.run()
+            rtime = time.time() - stime
+            print('finished in {0:.3f}s'.format(rtime))
 
         # Mapping level
-        print('\t{0:30s}\t'.format('Tags'), end='', flush=True)
-        cmd = ['donotremove', 
-               '-i', self.output_dir,
-               '-d', self.output_dir, 
-               '-l', 'mapping',
-               '-n', self.run_name]
-        if self.verbose:
-            print(' '.join(cmd))
-        self.update_status('Running {title} ({level})'.format(title='Aggregator', level='mapping'))
-        m_aggregator = Aggregator(cmd, self.status_writer)
-        m_aggregator.run()
-        rtime = time.time() - stime
-        print('finished in {0:.3f}s'.format(rtime))
+        if not self.append_mode:
+            print('\t{0:30s}\t'.format('Tags'), end='', flush=True)
+            cmd = ['donotremove', 
+                '-i', self.output_dir,
+                '-d', self.output_dir, 
+                '-l', 'mapping',
+                '-n', self.run_name]
+            if self.verbose:
+                print(' '.join(cmd))
+            self.update_status('Running {title} ({level})'.format(title='Aggregator', level='mapping'))
+            m_aggregator = Aggregator(cmd, self.status_writer)
+            m_aggregator.run()
+            rtime = time.time() - stime
+            print('finished in {0:.3f}s'.format(rtime))
 
         return v_aggregator.db_path
 
@@ -903,43 +966,48 @@ class Cravat (object):
         dbpath = os.path.join(self.output_dir, self.run_name + '.sqlite')
         conn = await aiosqlite3.connect(dbpath)
         cursor = await conn.cursor()
-        q = 'drop table if exists info'
-        await cursor.execute(q)
-        q = 'create table info (colkey text, colval text)'
-        await cursor.execute(q)
-        created = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        q = 'insert into info values ("Result created at", "' + created + '")'
-        await cursor.execute(q)
-        q = 'insert into info values ("Input file name", "{}")'.format(';'.join(self.inputs)) #todo adapt to multiple inputs
-        await cursor.execute(q)
-        q = 'insert into info values ("Input genome", "' + self.input_assembly + '")'
-        await cursor.execute(q)
-        q = 'select count(*) from variant'
-        await cursor.execute(q)
-        r = await cursor.fetchone()
-        no_input = str(r[0])
-        q = 'insert into info values ("Number of unique input variants", "' + no_input + '")'
-        await cursor.execute(q)
-        q = 'insert into info values ("open-cravat", "{}")'.format(self.pkg_ver)
-        await cursor.execute(q)
-        if hasattr(self, 'converter_format'):
-            q = 'insert into info values ("_converter_format", "{}")'.format(self.converter_format)
+        if not self.append_mode:
+            q = 'drop table if exists info'
             await cursor.execute(q)
-        if hasattr(self, 'genemapper'):
-            version = self.genemapper.conf['version']
-            title = self.genemapper.conf['title']
-            modulename = self.genemapper.name
-            genemapper_str = '{} ({})'.format(title, version)
-            q = 'insert into info values ("Gene mapper", "{}")'.format(genemapper_str)
+            q = 'create table info (colkey text primary key, colval text)'
             await cursor.execute(q)
-            q = 'insert into info values ("_mapper", "{}:{}")'.format(modulename, version)
+        modified = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        q = 'insert or replace into info values ("Result modified at", "' + modified + '")'
+        await cursor.execute(q)
+        if not self.append_mode:
+            created = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+            q = 'insert into info values ("Result created at", "' + created + '")'
             await cursor.execute(q)
-        f = open(os.path.join(self.output_dir, self.run_name + '.crm'))
-        for line in f:
-            if line.startswith('#input_paths='):
-                input_path_dict_str = '='.join(line.strip().split('=')[1:]).replace('"', "'")
-                q = 'insert into info values ("_input_paths", "{}")'.format(input_path_dict_str)
+            q = 'insert into info values ("Input file name", "{}")'.format(';'.join(self.inputs)) #todo adapt to multiple inputs
+            await cursor.execute(q)
+            q = 'insert into info values ("Input genome", "' + self.input_assembly + '")'
+            await cursor.execute(q)
+            q = 'select count(*) from variant'
+            await cursor.execute(q)
+            r = await cursor.fetchone()
+            no_input = str(r[0])
+            q = 'insert into info values ("Number of unique input variants", "' + no_input + '")'
+            await cursor.execute(q)
+            q = 'insert into info values ("open-cravat", "{}")'.format(self.pkg_ver)
+            await cursor.execute(q)
+            if hasattr(self, 'converter_format'):
+                q = 'insert into info values ("_converter_format", "{}")'.format(self.converter_format)
                 await cursor.execute(q)
+            if hasattr(self, 'genemapper'):
+                version = self.genemapper.conf['version']
+                title = self.genemapper.conf['title']
+                modulename = self.genemapper.name
+                genemapper_str = '{} ({})'.format(title, version)
+                q = 'insert into info values ("Gene mapper", "{}")'.format(genemapper_str)
+                await cursor.execute(q)
+                q = 'insert into info values ("_mapper", "{}:{}")'.format(modulename, version)
+                await cursor.execute(q)
+            f = open(os.path.join(self.output_dir, self.run_name + '.crm'))
+            for line in f:
+                if line.startswith('#input_paths='):
+                    input_path_dict_str = '='.join(line.strip().split('=')[1:]).replace('"', "'")
+                    q = 'insert into info values ("_input_paths", "{}")'.format(input_path_dict_str)
+                    await cursor.execute(q)
         q = 'select colval from info where colkey="annotators_desc"'
         await cursor.execute(q)
         r = await cursor.fetchone()
@@ -972,12 +1040,12 @@ class Cravat (object):
             module_info = au.get_local_module_info(name)
             if module_info is not None and module_info.conf is not None:
                 annotator_desc_dict[name] = module_info.conf['description']
-        q = 'insert into info values ("_annotator_desc", "{}")'.format(json.dumps(annotator_desc_dict).replace('"', "'"))
+        q = 'insert or replace into info values ("_annotator_desc", "{}")'.format(json.dumps(annotator_desc_dict).replace('"', "'"))
         await cursor.execute(q)
         self.status_writer.queue_status_update('annotator_version', annotator_version)
-        q = 'insert into info values ("Annotators", "' + annotators_str + '")'
+        q = 'insert or replace into info values ("Annotators", "' + annotators_str + '")'
         await cursor.execute(q)
-        q = 'insert into info values ("_annotators", "{}")'.format(','.join(annotators))
+        q = 'insert or replace into info values ("_annotators", "{}")'.format(','.join(annotators))
         await cursor.execute(q)
         await conn.commit()
         await cursor.close()
